streamtuner: Fix format warning with Qt 6
With Qt 5 QJsonArray.size() returns an int.
With Qt 6 it returns qsizetype aka long long int.

+++ src/streamtuner/shoutcast-model.cc
@@ -94,7 +94,7 @@ void ShoutcastTunerModel::process_station (QJsonObject object)
 
 void ShoutcastTunerModel::process_stations (QJsonArray & stations)
 {
-    AUDINFO ("Retrieved %d stations.\n", stations.size ());
+    AUDINFO ("Retrieved %lld stations.\n", (long long int) stations.size ());
 
     beginResetModel ();
